LP-437 Ensure empty string not sent for 'former_names' field
