+++ src/infrastructure/gateway/generalPartner/GeneralPartnerGateway.ts
@@ -6,7 +6,7 @@ import {
   GeneralPartner,
   LimitedPartnershipResourceCreated
 } from "@companieshouse/api-sdk-node/dist/services/limited-partnerships";
-import convertDateToIsoDateString from "../dateMapper";
+import { convertDateToIsoDateString, removeEmptyStringValues } from "../utils";
 
 class GeneralPartnerGateway implements IGeneralPartnerGateway {
   async createGeneralPartner(
@@ -14,6 +14,8 @@ class GeneralPartnerGateway implements IGeneralPartnerGateway {
     transactionId: string,
     data: Record<string, any>
   ): Promise<string> {
+    removeEmptyStringValues(data);
+
     if (data["forename"]) {
       // Only do this if General Partner Person data is being sent to the API
       data["date_of_birth"] = convertDateToIsoDateString(data["date_of_birth-day"], data["date_of_birth-month"], data["date_of_birth-year"]);
